volpath: correct attenuation for area lights surrounded by a medium
Previously, volpath did not correctly compute the attenuation to
an area light source when it was embedded inside a participating medium
(i.e. referencing it directly). volpath_simple was unaffected.

It is unlikely that anybody ran into this problem, since it was in fact
not possible to construct such a scene due to an assertion in shape.cpp.